+++ Sources/_FoundationInternals/String/BidirectionalCollection.swift
@@ -61,4 +61,72 @@ extension BidirectionalCollection {
         }
         return self[startOfNonTrimmedRange ... backIdx]
     }
+
+    // Equal to calling `index(&idx, offsetBy: -other.count)` with just one loop
+    func _index<S: BidirectionalCollection>(_ index: Index, backwardsOffsetByCountOf other: S) -> Index? {
+        var idx = index
+        var otherIdx = other.endIndex
+        while otherIdx > other.startIndex  {
+            guard idx > startIndex else {
+                // other.count > self.count: bail
+                return nil
+            }
+            other.formIndex(before: &otherIdx)
+            formIndex(before: &idx)
+        }
+        return idx
+    }
+
+    func _range<S: BidirectionalCollection>(of other: S, anchored: Bool = false, backwards: Bool = false) -> Range<Index>? where S.Element == Element, Element : Equatable {
+        var result: Range<Index>? = nil
+        var fromLoc: Index
+        var toLoc: Index
+        if backwards {
+            guard let idx = _index(endIndex, backwardsOffsetByCountOf: other) else {
+                // other.count > string.count: bail
+                return nil
+            }
+            fromLoc = idx
+
+            toLoc = anchored ? fromLoc : startIndex
+        } else {
+            fromLoc = startIndex
+            if anchored {
+                toLoc = fromLoc
+            } else {
+                guard let idx = _index(endIndex, backwardsOffsetByCountOf: other) else {
+                    return nil
+                }
+                toLoc = idx
+            }
+        }
+
+        let delta = fromLoc <= toLoc ? 1 : -1
+
+        while true {
+            var str1Index = fromLoc
+            var str2Index = other.startIndex
+
+            while str2Index < other.endIndex && str1Index < endIndex {
+                if self[str1Index] != other[str2Index] {
+                    break
+                }
+                formIndex(after: &str1Index)
+                other.formIndex(after: &str2Index)
+            }
+
+            if str2Index == other.endIndex {
+                result = fromLoc..<str1Index
+                break
+            }
+
+            if fromLoc == toLoc {
+                break
+            }
+
+            formIndex(&fromLoc, offsetBy: delta)
+        }
+
+        return result
+    }
 }
